+++ app/code/ThirdParty/BlogArticle/Block/Adminhtml/Category.php
@@ -16,7 +16,7 @@ class Category extends Template
     /**
      * @var FormKey
      */
-    private $formKey;
+    protected $formKey;
 
     public function __construct(
         Context $context,